requirements.txt
anthropic>=0.39.0
openai>=1.0.0
google-generativeai>=0.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-docx>=1.1.0
markdown>=3.5.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
